+++ application/models/Supplies.php
@@ -1,10 +1,7 @@
 <?php
 
 /**
- * This is a "CMS" model for quotes, but with bogus hard-coded data,
- * so that we don't have to worry about any database setup.
- * This would be considered a "mock database" model.
- *
+ * Supplies model with CRUD functions to interact with a backend database through REST calls
  * @author Gerard
  */
 define('REST_SERVER', 'http://backend.local');      // the REST server host
@@ -18,13 +15,25 @@ class Supplies extends MY_Model {
         //*** Explicitly load the REST libraries.
         $this->load->library(['curl', 'format', 'rest']);
     }
-
+    /**
+     * Returns all the ports from the xml file
+     * @return the ports
+     */
+    function getPorts()
+    {
+        $ports = array();
+        foreach ($this->xml->ports->children() as $port)
+        {
+            $ports[(string) $port['code']] = $port->__toString();
+        }
+        return $ports;
+    }
     function rules() {
         $config = [
             ['field'=>'id', 'label'=>'Menu code'],
             ['field'=>'name', 'label'=>'Item name'],
             ['field'=>'qty_onhand', 'label'=>'Item onhand'],
-            ['field'=>'qty_inventory', 'label'=>'Item stock', 'rules'=> 'required|decimal'],
+            ['field'=>'qty_inventory', 'label'=>'Item stock', 'rules'=> 'required'],
             ['field'=>'price', 'label'=>'Price'],
         ];
         return $config;
@@ -85,7 +94,8 @@ class Supplies extends MY_Model {
     {
         $this->rest->initialize(array('server' => REST_SERVER));
         $this->rest->option(CURLOPT_PORT, REST_PORT);
-        $retrieved = $this->rest->put('/maintenance/item/id/' . $record->id, $record);
+
+        $retrieved = $this->rest->put('/maintenance/item/id/' . $record->id, json_encode($record));
     }
 
     // Add a record to the DB
@@ -93,6 +103,6 @@ class Supplies extends MY_Model {
     {
         $this->rest->initialize(array('server' => REST_SERVER));
         $this->rest->option(CURLOPT_PORT, REST_PORT);
-        $retrieved = $this->rest->post('/maintenance/item/id/' . $record->id, $record);
+        $retrieved =  $this->rest->post('/maintenance/item/id/' . $record->id, json_encode($record));
     }
 }