+++ application/views/administrator_view.php
@@ -2,6 +2,10 @@
     <button type="button" class="close" data-dismiss="alert" aria-label="Close"><span aria-hidden="true">&times;</span></button>
     <strong>Hey!</strong> You can minimize the tables by clicking on the headers.
 </div>
+<div class="alert alert-success alert-dismissible" role="alert">
+    <button type="button" class="close" data-dismiss="alert" aria-label="Close"><span aria-hidden="true">&times;</span></button>
+    To prevent system overload, any price adjustments are rounded down to the nearest integer.
+</div>
 <div class="panel-group" id="accordion" role="tablist" aria-multiselectable="true">
     <div class="panel panel-default">
         <div class="panel-heading" role="tab" id="headingOne">
@@ -16,11 +20,11 @@
 
                 <table class = "table table-bordered">
                     <tr>
-                        <th>Id</th>
-                        <th>Name</th>
-                        <th>On-Hand Quantity</th>
-                        <th>Receiving Quantity</th>
-                        <th>Price</th>
+                        <th>Item Id</th>
+                        <th>Item Name</th>
+                        <th>On-Hand Quantity (g)</th>
+                        <th>Receiving Quantity (g)</th>
+                        <th>Price per Unit ($)</th>
                     </tr>
                     {supplyItems}
                     <tr> <td><a class="btn btn-default" role="button" href="/Administrator/editSupplies/{id}">{id}</a></td>
@@ -58,11 +62,11 @@
                         <th>Item Name</th>
                         <th>Description</th>
                         <th>Price</th>
-                        <th>Units</th>
+                        <th>Units Available (Pieces)</th>
                         <th>Picture</th>
                     </tr>
                     {recipeItems}
-                    <tr> <td><a class="btn btn-default" role="button" href="/Administrator/editRecipes/{id}">{id}</a></td>
+                    <tr> <td><a class="btn btn-default" role="button" href="/Production/edit/{id}">{id}</a></td>
                         <td>{name}</td>
                         <td>{description}</td>
                         <td>{price}</td>