+++ application/views/sales-order_view.php
@@ -1,8 +1,8 @@
-<div class="col-sm-6 col-md-3">
+<div class="col-sm-6 col-md-3 col-lg-3">
     <div class="thumbnail">
       <a href="/Sales/add/{id}"><img class="scale" src="/pix/{picture}"/></a><br/>
         <h5 class="text-center">{name}</h5>
         <h5 class="text-center">Price: ${price}</h5>
-        <h5 class="text-center">Units Available: {unit}</h5>
+        <h5 class="text-center">Units Available: {unit} Pieces</h5>
     </div>
 </div>