+++ changelog.md
@@ -7,8 +7,18 @@ Team members:
 @gerardng
 
 ## [2.0.0] - 2016-12-11
+### Changed 2016-12-10
+- Receiving page forms properly adding up to inventory quantity @gerardng
+- Added prepare function to convert receiving inventory into on-hand quantity @gerardng
+- Rewrite the production feature @AlexandyZ
+
+### Changed 2016-12-09
+- fixed Receiving page properly connecting to Supplies table in backend DB @gerardng
+- updated database values to align with units @New3
+
 ### Changed - 2016-12-08
 - redesigned Home page to track capital, total sales, ingredients consumed, receiving orders @gerardng
+- Delete the productions model and create Recipe_supply model instead @AlexandyZ
 
 ### Added - 2016-12-07
 - Separated front & backend features @AlexandyZ @New3